pyproject.toml
[tool.poetry]
name = "modbot"
version = "0.1.0"
description = "The moderation bot used within the Créu Cat discord server"
authors = ["Sofia \"dzshn\" N. L. <zshn@pm.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
